==> 02.0.compute_metrics/01.table_with_bed_clusters.ipynb <==
import numpy as np
import pandas as pd
import datetime
#import dcarte
import pytz
#%load_ext autoreload
import pyarrow as pa
import pyarrow.parquet as pq

from random import randrange
import matplotlib.pyplot as plt
import seaborn as sns
color = sns.color_palette()
pd.options.mode.chained_assignment = None  # default='warn'

sleep = pd.read_parquet('sleep_data.parquet', engine='pyarrow')
sleep = sleep.drop('duplicates', axis=1)
ids_list = sleep['ID'].unique()

len(ids_list)

def process_sleep_mat(patient_id, df):
    df = df[df['ID']==patient_id].reset_index(drop=True)

    df.index = pd.DatetimeIndex(df['start_date_tz'])
    df = df.drop(columns=['start_date_tz'])
    df = df.resample('1T').asfreq()#.ffill()
    df = df.rename_axis('datetime').reset_index()
    df['value'] = np.where(df['state'].isna(), 'bed_out', 'bed_in')
    df['value_diff'] = df['value']!=df['value'].shift(1)
    df['ID'] = patient_id
    return df[df['value_diff']==True][['ID','datetime','value']].reset_index(drop=True)

def start_table():
    return pd.DataFrame(columns=['Participant ID','date','cluster','datetime','end_datetime','length','classification'])
def append_row(df, row):
    return pd.concat([ df, pd.DataFrame([row], columns=row.index)]).reset_index(drop=True)
def get_date(df):
        df['hour'] = df['datetime'].dt.hour
        df['date'] = df['datetime'].dt.date
        df['delta'] = pd.Timedelta(days=0)
        df.loc[df['hour']<7, 'delta'] = pd.Timedelta(days=1)
        df['date'] = df['date']-df['delta']
        return df[['date','datetime', 'end_datetime','length']]

table=start_table()
for i in ids_list:
    if i!='Xv41A':
        i_df = process_sleep_mat(i, sleep)
        i_df['duration_in_minutes'] = (i_df['datetime'].shift(-1)-i_df['datetime']).dt.total_seconds()/60
        i_df['datetime'] = i_df['datetime'].dt.tz_convert(pytz.timezone('Europe/London'))
        i_df['end_datetime'] = i_df['datetime'].shift(-1)
        i_df['duration_in_minutes'] = (i_df['end_datetime']-i_df['datetime']).dt.total_seconds()/60
        i_df = i_df[i_df['value']=='bed_in']
        i_df = i_df.assign(gap  = (i_df['datetime'] - i_df['end_datetime'].shift()).dt.total_seconds()/60 )
        i_df = i_df.assign(cluster  = ((i_df['gap'] > 60).cumsum()))
        group = i_df.groupby(['cluster']).aggregate({'datetime':min,'end_datetime':max})
        group['length'] = (group['end_datetime']-group['datetime']).dt.total_seconds()/60
        group = get_date(group)
        group['Participant ID'] = i
        group['classification'] = 'diurnal'
        group.loc[(group['datetime'].dt.hour>=18)|(group['datetime'].dt.hour<7), 'classification'] = 'nocturnal'


        table  = pd.concat([ table, group.reset_index()]).reset_index(drop=True)
        df2 = pa.Table.from_pandas(table)
        pq.write_table(df2, 'clusters_hour_gap.parquet', compression='BROTLI')

table

table = pd.read_parquet('clusters_hour_gap.parquet', engine='pyarrow')

table['time'] = table['datetime'].dt.hour

sns.histplot(data=table, x='time',hue='classification',binwidth=1)

table.sort_values(by=['length'])

table[(table['time']<7)&(table['time']>5)]
